### 2020-05-22 #2220 @whedon
[PRE REVIEW]: GRLC - the Git Repository Linked data api Constructor
**Submitting author:** @c-martinez (<a href="http://orcid.org/0000-0001-5565-7577">Carlos Martinez-Ortiz</a>)
**Repository:** <a href="https://github.com/CLARIAH/grlc" target ="_blank">https://github.com/CLARIAH/grlc</a>
**Version:** v1.3.4
