### 2020-05-22 #2220 @arfon
:wave: @c-martinez. Thanks for your submission to JOSS. As described in [our blog post](https://blog.joss.theoj.org/2020/05/reopening-joss) announcing the reopening of JOSS, we're currently working in a "reduced service mode", limiting the number of papers assigned to any individual editor.

Thanks in advance for your patience!
