### 2020-06-11 73ed415
automate version checks
