### 2020-06-11 db114e0
Test grlc as library
