### 2020-06-15 #2220 @lorenanicole
@JanWielemaker @essepuntato just wanted to check in to see if there was anything further you may need?

> Greetings @JanWielemaker @essepuntato! Would you be willing to review this submission for JOSS?
>
> We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines 👉 https://joss.readthedocs.io/en/latest/reviewer_guidelines.html.
