### 2020-06-15 #2220 @essepuntato
> @JanWielemaker @essepuntato just wanted to check in to see if there was anything further you may need?

Hi @lorenanicole, apologies – I'm just back from holidays right now. I'll work on as soon as possible. Would it be possible for me to work on it next week or it is too late?

Have a nice day :-)
