### 2020-06-16 dd8cfc3
typo fixes
ssuming
synax
