### 2020-06-17 #2220 @lorenanicole
@essepuntato not too late at all! Just wanted to circle back in case we needed to find another editor. Thank you for following up! This comment will be of use to you 👇  https://github.com/openjournals/joss-reviews/issues/2220#issuecomment-641390482
