### 2020-06-25 #2220 @essepuntato
> @essepuntato not too late at all! Just wanted to circle back in case we needed to find another editor. Thank you for following up! This comment will be of use to you 👇 [#2220 (comment)](https://github.com/openjournals/joss-reviews/issues/2220#issuecomment-641390482)

Hi @lorenanicole

it seems I forgot to formally accept the invitation by whedon and it expired. Could you please sent it again?
