### 2020-06-25 b2299c3
Update workflow to tweet only on releases
