+++ .github/workflows/tweet-release.yml
@@ -1,34 +1,23 @@
-# This is a basic workflow to help you get started with Actions
 name: Tweet release
-
-# Controls when the action will run. Triggers the workflow on push or pull request
-# events but only for the master branch
+# Tweets will happen when a new release is published
 on:
-  push:
-    branches: [ gh-actions ]
   release:
-    types: 
+    types:
       - published
-
-# A workflow run is made up of one or more jobs that can run sequentially or in parallel
 jobs:
-  # This workflow contains a single job called "build"
   tweet:
-    # The type of runner that the job will run on
     runs-on: ubuntu-latest
-    # Steps represent a sequence of tasks that will be executed as part of the job
     steps:
     - name: GitHub Twittction
       uses: m1ner79/Github-Twittction@v1.0.1
       with:
         # Message posted on twitter
-        # twitter_status: "Release ${{github.event.release.tag_name}} of ${{github.event.repository.full_name}} is out! Check out the release notes ${{github.event.release.html_url}}"
-        twitter_status: "Hello twitter!"
+        twitter_status: "Release ${{github.event.release.tag_name}} of ${{github.event.repository.full_name}} is out! Check out the release notes ${{github.event.release.html_url}}"
         # Twitter Consumer Key is referenced here
-        twitter_consumer_key: ${{ secrets.TWITTER_CONSUMER_KEY }} 
+        twitter_consumer_key: ${{ secrets.TWITTER_CONSUMER_KEY }}
         # Twitter Consumer Secret is referenced here
-        twitter_consumer_secret: ${{ secrets.TWITTER_CONSUMER_SECRET }} 
+        twitter_consumer_secret: ${{ secrets.TWITTER_CONSUMER_SECRET }}
         # Twitter Access Token Key is referenced here
-        twitter_access_token_key: ${{ secrets.TWITTER_ACCESS_TOKEN_KEY }} 
+        twitter_access_token_key: ${{ secrets.TWITTER_ACCESS_TOKEN_KEY }}
         # Twitter Access Token Secret is referenced here
-        twitter_access_token_secret: ${{ secrets.TWITTER_ACCESS_TOKEN_SECRET }} 
+        twitter_access_token_secret: ${{ secrets.TWITTER_ACCESS_TOKEN_SECRET }}
