### 2020-06-26 #2220 @lorenanicole
Hi 👋 @jonathanblaney, would you be willing to review this submission for JOSS?

We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines 👉  https://joss.readthedocs.io/en/latest/reviewer_guidelines.html
