### 2020-07-29 d2143b5
Add tests for API paths
