### 2020-07-29 c8850c0
Backward compatibility of /spec endpoints
