+++ src/server.py
@@ -81,6 +81,8 @@ def api_docs_local():
 # Spec generation, JSON
 @app.route('/api-local/swagger', methods=['GET'])
 @app.route('/api/local/local/swagger', methods=['GET'], strict_slashes=False)  # backward compatibility route
+@app.route('/api-local/spec', methods=['GET'])                              # backward compatibility route
+@app.route('/api/local/local/spec', methods=['GET'], strict_slashes=False)  # backward compatibility route
 def swagger_spec_local():
     """Swagger spec for local routes."""
     return swagger_spec(user=None, repo=None, sha=None, content=None)
@@ -107,6 +109,7 @@ def api_docs_param():
 
 # Spec generation, JSON
 @app.route('/api-url/swagger', methods=['GET'])
+@app.route('/api-url/spec', methods=['GET'])      # backward compatibility route
 def swagger_spec_param():
     """Swagger spec for specifications loaded via http."""
     spec_url = request.args['specUrl']
@@ -140,7 +143,7 @@ def query_param(query_name):
 @app.route('/api/<user>/<repo>/commit/<sha>/api-docs')  # backward compatibility route
 @app.route('/api/<user>/<repo>/<subdir>/commit/<sha>')  # backward compatibility route
 @app.route('/api/<user>/<repo>/<subdir>/commit/<sha>/api-docs')  # backward compatibility route
-def api_docs_git(user, repo, subdir=None, spec_url=None, sha=None):
+def api_docs_git(user, repo, subdir=None, sha=None):
     """Grlc API page for specifications loaded from a Github repo."""
     return api_docs_template()
 
@@ -150,13 +153,22 @@ def api_docs_git(user, repo, subdir=None, spec_url=None, sha=None):
 @app.route('/api-git/<user>/<repo>/commit/<sha>/swagger')
 @app.route('/api-git/<user>/<repo>/subdir/<subdir>/commit/<sha>/swagger')
 @app.route('/api-git/<user>/<repo>/<subdir>/commit/<sha>/swagger')
-@app.route('/api-git/<user>/<repo>/swagger', methods=['GET'])  # backward compatibility route
-@app.route('/api-git/<user>/<repo>/<subdir>/swagger', methods=['GET'])  # backward compatibility route
-@app.route('/api-git/<user>/<repo>/commit/<sha>/swagger')  # backward compatibility route
-@app.route('/api-git/<user>/<repo>/<subdir>/commit/<sha>/swagger')  # backward compatibility route
-def swagger_spec_git(user, repo, subdir=None, spec_url=None, sha=None, content=None):
+@app.route('/api/<user>/<repo>/swagger', methods=['GET'])  # backward compatibility route
+@app.route('/api/<user>/<repo>/<subdir>/swagger', methods=['GET'])  # backward compatibility route
+@app.route('/api/<user>/<repo>/commit/<sha>/swagger')  # backward compatibility route
+@app.route('/api/<user>/<repo>/<subdir>/commit/<sha>/swagger')  # backward compatibility route
+@app.route('/api-git/<user>/<repo>/spec', methods=['GET'])  # backward compatibility route
+@app.route('/api-git/<user>/<repo>/subdir/<subdir>/spec', methods=['GET'])  # backward compatibility route
+@app.route('/api-git/<user>/<repo>/commit/<sha>/spec')  # backward compatibility route
+@app.route('/api-git/<user>/<repo>/subdir/<subdir>/commit/<sha>/spec')  # backward compatibility route
+@app.route('/api-git/<user>/<repo>/<subdir>/commit/<sha>/spec')  # backward compatibility route
+@app.route('/api/<user>/<repo>/spec', methods=['GET'])  # backward compatibility route
[…]
