### 2020-08-13 #2220 @lorenanicole
Hi @alexdma, circling back! would you be willing to review this submission for JOSS?
