### 2020-08-18 #2220 @alexdma
Hi @lorenanicole - yes, sorry for the delay in getting back to you, but I still do not quite understand where and how I am supposed to write my review. Do I have to open an issue here, comment on this one or... what? And where do I go to issue commands to whedon?
