### 2020-09-02 d09cf02
Check for debug in docker image
