### 2020-09-03 0fc624d
Version management in a single file (CITATION.CFF)
