### 2020-09-07 3464528
Fix bug on <subdir> execution path building
