### 2020-09-27 #2220 @labarba
hi 👋  – @lorenanicole: it looks like the reviewer needs some guidance here. Can you pop in to help?

@openjournals/dev – can you check why a reviewer is listed twice on the header comment? Should this be edited manually?
