### 2020-09-27 #2220 @xuanxu
> can you check why a reviewer is listed twice on the header comment? Should this be edited manually?

He was added twice (in june [17th](https://github.com/openjournals/joss-reviews/issues/2220#issuecomment-645573740) and [26th](https://github.com/openjournals/joss-reviews/issues/2220#issuecomment-650300085)). I have edited it manually.
