### 2020-09-27 a98aed4
Adding support for customizable local repo configuration
