+++ src/fileLoaders.py
@@ -1,5 +1,6 @@
 import grlc.static as static
 from grlc.queryTypes import qType
+import grlc.glogging as glogging
 
 import json
 import requests
@@ -8,7 +9,9 @@ from os import path
 from glob import glob
 from github import Github
 from github.GithubException import BadCredentialsException
+from configparser import ConfigParser
 
+glogger = glogging.getGrlcLogger(__name__)
 
 class BaseLoader:
     """Base class for File Loaders"""
@@ -39,13 +42,6 @@ class BaseLoader:
                 return queryText, queryType
         # No query found...
         return '', None
-
-    def getLicenceURL(self):
-        """Returns the URL of the license file in this repository if one exists."""
-        for f in self.fetchFiles():
-            if f['name'].lower() == 'license' or f['name'].lower() == 'licence':
-                return f['download_url']
-        return None
 
     def _getText(self, queryFullName):
         """To be implemented by sub-classes"""
@@ -155,6 +151,13 @@ class GithubLoader(BaseLoader):
         """Return the full URI of the github repo."""
         return static.GITHUB_API_BASE_URL + self.gh_repo.full_name
 
+    def getLicenceURL(self):
+        """Returns the URL of the license file in this repository if one exists."""
+        for f in self.fetchFiles():
+            if f['name'].lower() == 'license' or f['name'].lower() == 'licence':
+                return f['download_url']
+        return None
+
 
 class LocalLoader(BaseLoader):
     """Local file system file loader. Retrieves information to construct 
@@ -166,6 +169,21 @@ class LocalLoader(BaseLoader):
         Keyword arguments:
         baseDir -- Local file system path where files are loaded from."""
         self.baseDir = baseDir
+
+        config_fallbacks = {
+            'repo_title': 'local',
+            'contact_name': '',
+            'contact_url': '',
+            'licence_url': ''
+        }
+        config = ConfigParser(config_fallbacks)
+        config.add_section('repo_info')
+        config_filename = path.join(baseDir, 'local-repo-config.ini')
+        config.read(config_filename)
+        self.repo_title = config.get('repo_info', 'repo_title')
+        self.contact_name = config.get('repo_info', 'contact_name')
+        self.contact_url = config.get('repo_info', 'contact_url')
+        self.licence_url = config.get('repo_info', 'licence_url')
 
     def fetchFiles(self):
         """Returns a list of file items contained on the local repo."""
@@ -202,15 +220,15 @@ class LocalLoader(BaseLoader):
 
     def getRepoTitle(self):
         """Return the title of the local repo."""
-        return 'local'
+        return self.repo_title
 
     def getContactName(self):
         """Return the name of the owner of the local repo."""
-        return ''
+        return self.contact_name
 
     def getContactUrl(self):
         """Return the home page of the owner of the local repo."""
-        return ''
+        return self.contact_url
 
     def getCommitList(self):
         """Return a list of commits (always a single commit) on the local repo."""
@@ -224,6 +242,8 @@ class LocalLoader(BaseLoader):
         """Return the full URI of the local repo."""
         return 'local-file-system'
 
+    def getLicenceURL(self):
+        return self.licence_url
 
 class URLLoader(BaseLoader):
     """URL specification loader. Retrieves information to construct a grlc 
