+++ src/fileLoaders.py
@@ -158,6 +158,10 @@ class GithubLoader(BaseLoader):
                 return f['download_url']
         return None
 
+    def getRepoDescription(self):
+        """Return the description of the repository"""
+        return self.gh_repo.description
+
 
 class LocalLoader(BaseLoader):
     """Local file system file loader. Retrieves information to construct 
@@ -172,15 +176,17 @@ class LocalLoader(BaseLoader):
 
         config_fallbacks = {
             'repo_title': 'local',
+            'api_description': 'API generated from local files',
             'contact_name': '',
             'contact_url': '',
             'licence_url': ''
         }
         config = ConfigParser(config_fallbacks)
         config.add_section('repo_info')
-        config_filename = path.join(baseDir, 'local-repo-config.ini')
+        config_filename = path.join(baseDir, 'local-api-config.ini')
         config.read(config_filename)
         self.repo_title = config.get('repo_info', 'repo_title')
+        self.api_description = config.get('repo_info', 'api_description')
         self.contact_name = config.get('repo_info', 'contact_name')
         self.contact_url = config.get('repo_info', 'contact_url')
         self.licence_url = config.get('repo_info', 'licence_url')
@@ -244,6 +250,11 @@ class LocalLoader(BaseLoader):
 
     def getLicenceURL(self):
         return self.licence_url
+
+    def getRepoDescription(self):
+        """Return the description of the API generated by local repository"""
+        return self.api_description
+
 
 class URLLoader(BaseLoader):
     """URL specification loader. Retrieves information to construct a grlc 
@@ -330,3 +341,11 @@ class URLLoader(BaseLoader):
     def getLicenceURL(self):
         """Returns the URL of the license file in the specification."""
         return self.spec['licence'] if self.spec['licence'] else ''
+
+    def getRepoDescription(self):
+        """Return the description of the repository"""
+        if 'description' in self.spec:
+            return self.spec['description']
+        else:
+            return 'API definition loaded from ' + self.getRawRepoUri()
+
