### 2020-10-05 0fbc8d5
Support multiple subdirs when using <subdir> option
