+++ src/server.py
@@ -127,33 +127,33 @@ def query_param(query_name):
 
 # Spec generation, front-end
 @app.route('/api-git/<user>/<repo>', strict_slashes=False)
-@app.route('/api-git/<user>/<repo>/subdir/<subdir>', strict_slashes=False)
+@app.route('/api-git/<user>/<repo>/subdir/<path:subdir>', strict_slashes=False)
 @app.route('/api-git/<user>/<repo>/api-docs')
 @app.route('/api-git/<user>/<repo>/commit/<sha>')
 @app.route('/api-git/<user>/<repo>/commit/<sha>/api-docs')
-@app.route('/api-git/<user>/<repo>/subdir/<subdir>/commit/<sha>')
-@app.route('/api-git/<user>/<repo>/subdir/<subdir>/commit/<sha>/api-docs')
+@app.route('/api-git/<user>/<repo>/subdir/<path:subdir>/commit/<sha>')
+@app.route('/api-git/<user>/<repo>/subdir/<path:subdir>/commit/<sha>/api-docs')
 @app.route('/api/<user>/<repo>', strict_slashes=False)  # backward compatibility route
-@app.route('/api/<user>/<repo>/<subdir>', strict_slashes=False)  # backward compatibility route
+@app.route('/api/<user>/<repo>/<path:subdir>', strict_slashes=False)  # backward compatibility route
 @app.route('/api/<user>/<repo>/api-docs')  # backward compatibility route
 @app.route('/api/<user>/<repo>/commit/<sha>')  # backward compatibility route
 @app.route('/api/<user>/<repo>/commit/<sha>/api-docs')  # backward compatibility route
-@app.route('/api/<user>/<repo>/<subdir>/commit/<sha>')  # backward compatibility route
-@app.route('/api/<user>/<repo>/<subdir>/commit/<sha>/api-docs')  # backward compatibility route
+@app.route('/api/<user>/<repo>/<path:subdir>/commit/<sha>')  # backward compatibility route
+@app.route('/api/<user>/<repo>/<path:subdir>/commit/<sha>/api-docs')  # backward compatibility route
 def api_docs_git(user, repo, subdir=None, sha=None):
     """Grlc API page for specifications loaded from a Github repo."""
     return api_docs_template()
 
 # Spec generation, JSON
 @app.route('/api-git/<user>/<repo>/swagger', methods=['GET'])
-@app.route('/api-git/<user>/<repo>/subdir/<subdir>/swagger', methods=['GET'])
+@app.route('/api-git/<user>/<repo>/subdir/<path:subdir>/swagger', methods=['GET'])
 @app.route('/api-git/<user>/<repo>/commit/<sha>/swagger')
-@app.route('/api-git/<user>/<repo>/subdir/<subdir>/commit/<sha>/swagger')
-@app.route('/api-git/<user>/<repo>/<subdir>/commit/<sha>/swagger')
+@app.route('/api-git/<user>/<repo>/subdir/<path:subdir>/commit/<sha>/swagger')
+@app.route('/api-git/<user>/<repo>/<path:subdir>/commit/<sha>/swagger')
 @app.route('/api-git/<user>/<repo>/swagger', methods=['GET'])  # backward compatibility route
-@app.route('/api-git/<user>/<repo>/<subdir>/swagger', methods=['GET'])  # backward compatibility route
+@app.route('/api-git/<user>/<repo>/<path:subdir>/swagger', methods=['GET'])  # backward compatibility route
 @app.route('/api-git/<user>/<repo>/commit/<sha>/swagger')  # backward compatibility route
-@app.route('/api-git/<user>/<repo>/<subdir>/commit/<sha>/swagger')  # backward compatibility route
[…]
