### 2020-10-05 3262390
Add tests for different endpoints
