### 2020-10-07 fc9545d
Fix endpoint testing for python3.7
