### 2020-10-07 #2220 @lorenanicole
> @alexdma We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines 👉 https://joss.readthedocs.io/en/latest/reviewer_guidelines.html. If you can please confirm, after reviewing these guidelines, that you are prepared to be a reviewer I can move us to the next step. We will open a separate issue for review.

Greetings @alexdma just checking in to see if you needed help with the above?
