### 2020-10-07 #2220 @alexdma
@lorenanicole thanks, but much as I may go over the guidelines, I cannot find in them the answer to my fundamental question.

_Where do I find the review page with the checklist that I have to fill?_

Here? I don't see the space to enter the review
On a new issue of this project? If I try to open one here, I get an automated message saying I must not create issues here.
I open an issue on the main GRLC repository? I have searched older issues there, and none of them seems to be a review.

So... where is the space for me to enter my review?

Thanks
