### 2020-10-07 #2220 @lorenanicole
@alexdma your review will have on a separate GitHub issues upon confirming your intent to review, basically all I need is a "yes I will review" from you. Then we'll use our JOSS tooling to open the review issue for your work to start! An example of a review issue is here: https://github.com/openjournals/joss-reviews/issues/2727.

Does that help answer your question?
