### 2020-10-12 446a544
Show build_spec warnings on UI
