### 2020-10-25 22b4e53
Replace missing PyGithub calls
