+++ tests/mock_data.py
@@ -2,34 +2,33 @@ from mock import Mock
 from os import path
 from glob import glob
 
+from collections import namedtuple
+
 base_url = path.join('tests', 'repo')
 def buildEntry(entryName):
     entryName = entryName.replace(base_url, '')
-    return {
-        u'download_url': entryName,
-         u'name': entryName,
-         u'path': entryName,
-         u'type': u'file'
-    }
+
+    # Named tuple containing properties of mocked github ContentFile
+    MockGithubContentFile = namedtuple('MockGithubContentFile', 'download_url name path type decoded_content')
+    return MockGithubContentFile(
+        download_url = entryName,
+        name = entryName,
+        path = entryName,
+        type = u'file',
+        decoded_content = 'FAKE FILE CONTENT'.encode() # Because Github ContentFile object contains bytes.
+    )
 mock_files = [ buildEntry(f) for f in glob(path.join(base_url, '*')) ]
 
-def mock_requestsGithub(uri, headers={}, params={}):
-    if uri.endswith('contents'):
-        return_value = Mock(ok=True)
-        return_value.json.return_value = mock_files
-        return return_value
-    else:
-        targetFile = uri.replace('https://raw.githubusercontent.com/fakeuser/fakerepo/master/', path.join(base_url, ''))
-        if path.exists(targetFile):
-            f = open(targetFile, 'r')
-            lines = f.readlines()
-            text = ''.join(lines)
-            return_value = Mock(status_code=200)
-            return_value.text = text
-            return return_value
+class MockGithubRepo:
+    def get_contents(self, filename, ref=None):
+        if filename == "":
+            return mock_files
         else:
-            return_value = Mock(status_code=404)
-            return return_value
+            for f in mock_files:
+                if filename in f.name: # filenames contain extra /
+                    return f
+            return None
+
 
 def mock_requestsUrl(url, headers={}, params={}):
     url = url.replace('http://example.org/', 'tests/repo/')
@@ -61,10 +60,10 @@ def mock_process_sparql_query_text(query_text, raw_repo_uri, call_name, extraMet
     }
     return mockItem
 
-
 filesInRepo = [
     {
         u'name': u'fakeFile1.rq',
         u'download_url': u'https://example.org/path/to/fakeFile.rq',
+        u'decoded_content': 'CONTENT ?'.encode() # Because Github ContentFile object contains bytes.
     }
 ]
