### 2020-11-23 dd46ba0
Update documentation for 'defaults' decorator
