### 2020-12-26 19a5910
Remove support for /spec in favour of /swagger
