+++ src/server.py
@@ -81,8 +81,6 @@ def api_docs_local():
 # Spec generation, JSON
 @app.route('/api-local/swagger', methods=['GET'])
 @app.route('/api/local/local/swagger', methods=['GET'], strict_slashes=False)  # backward compatibility route
-@app.route('/api-local/spec', methods=['GET'])                              # backward compatibility route
-@app.route('/api/local/local/spec', methods=['GET'], strict_slashes=False)  # backward compatibility route
 def swagger_spec_local():
     """Swagger spec for local routes."""
     return swagger_spec(user=None, repo=None, sha=None, content=None)
@@ -109,7 +107,6 @@ def api_docs_param():
 
 # Spec generation, JSON
 @app.route('/api-url/swagger', methods=['GET'])
-@app.route('/api-url/spec', methods=['GET'])      # backward compatibility route
 def swagger_spec_param():
     """Swagger spec for specifications loaded via http."""
     spec_url = request.args['specUrl']
@@ -157,18 +154,8 @@ def api_docs_git(user, repo, subdir=None, sha=None):
 @app.route('/api/<user>/<repo>/<subdir>/swagger', methods=['GET'])  # backward compatibility route
 @app.route('/api/<user>/<repo>/commit/<sha>/swagger')  # backward compatibility route
 @app.route('/api/<user>/<repo>/<subdir>/commit/<sha>/swagger')  # backward compatibility route
-@app.route('/api-git/<user>/<repo>/spec', methods=['GET'])  # backward compatibility route
-@app.route('/api-git/<user>/<repo>/swagger', methods=['GET'])  # backward compatibility route
-@app.route('/api-git/<user>/<repo>/subdir/<subdir>/spec', methods=['GET'])  # backward compatibility route
 @app.route('/api-git/<user>/<repo>/<path:subdir>/swagger', methods=['GET'])  # backward compatibility route
-@app.route('/api-git/<user>/<repo>/commit/<sha>/spec')  # backward compatibility route
-@app.route('/api-git/<user>/<repo>/subdir/<subdir>/commit/<sha>/spec')  # backward compatibility route
-@app.route('/api-git/<user>/<repo>/<subdir>/commit/<sha>/spec')  # backward compatibility route
 @app.route('/api-git/<user>/<repo>/<path:subdir>/commit/<sha>/swagger')  # backward compatibility route
-@app.route('/api/<user>/<repo>/spec', methods=['GET'])  # backward compatibility route
-@app.route('/api/<user>/<repo>/<subdir>/spec', methods=['GET'])  # backward compatibility route
-@app.route('/api/<user>/<repo>/commit/<sha>/spec')  # backward compatibility route
-@app.route('/api/<user>/<repo>/<subdir>/commit/<sha>/spec')  # backward compatibility route
 def swagger_spec_git(user, repo, subdir=None, sha=None):
     """Swagger spec for specifications loaded from a Github repo."""
     return swagger_spec(user, repo, subdir=subdir, spec_url=None, sha=sha, content=None)
