### 2021-01-11 #2731 @alexdma
My apologies @arfon - it is certainly no justification that I have moved to a new job and country in the meantime.

I'm trying to work through it but I can't seem to be able to tick any of the checkboxes on my list (code of conduct etc.) - how do I do that? The issue does not seem to be editable.
