### 2021-01-11 #2731 @arfon
> I'm trying to work through it but I can't seem to be able to tick any of the checkboxes on my list (code of conduct etc.) - how do I do that? The issue does not seem to be editable.

Thanks for the update @alexdma. Please accept the invite from @whedon above and then you should be able to update the checklist.
