### 2021-01-25 #2731 @alexdma
Hi - Does the Softyware paper section of the review checklist refer to the article as found here: https://github.com/openjournals/joss-papers/blob/joss.02731/joss.02731/10.21105.joss.02731.pdf ?

Thanks
