### 2021-01-25 CLARIAH/grlc#311 @alexdma
Document support for Linked Data Fragments
The [feature list](https://github.com/CLARIAH/grlc/wiki/Features) mentions compatibility of GRLC with LD fragments.

Whether this means that one can supply an LD fragment server as an `endpoint` parameter, then a specification as to what the query should look like, in order to be parsable by GRLC, would make things clearer. Or even if I have completely mistaken the notion of LDF compatibility for GRLC, it would best be documented.

It would be useful if a small tutorial could be provided in the documentation, with examples running on http://fragments.dbpedia.org or others.

Thanks!
