### 2021-01-26 CLARIAH/grlc#311 @c-martinez
Hi @alexdma, Thank you for your comment. I suppose this is related to https://github.com/openjournals/joss-reviews/issues/2731, correct?

Indeed it would be good to have a well documented example of how to use this feature. Perhaps as part of the [quick tutorial](https://github.com/CLARIAH/grlc/wiki/Quick-tutorial).

@albertmeronyo -- what do you think? Do you have an old example we could use to document this feature?
