### 2021-01-27 CLARIAH/grlc#311 @alexdma
Hi @c-martinez - that is correct, I will be sure to mention the JOSS ticket in any related issues I might open.
