### 2021-02-05 858959d
Restrict transform decorator to application/json
