### 2021-02-05 7447532
Catching exception connecting to SPARQL endpoint
