### 2021-02-17 #2731 @lorenanicole
@essepuntato just bumping the below to see if you needed anything additional:

https://github.com/openjournals/joss-reviews/issues/2731#issuecomment-774939456
