### 2021-03-26 CLARIAH/grlc#325 @alexdma
[Software paper] Describe the state of the field
Re openjournals/joss-reviews#2731

Paragraph 2 of the JOSS paper at https://github.com/openjournals/joss-papers/blob/joss.02731/joss.02731/10.21105.joss.02731.pdf

describes the motivation of Grlc properly. It would best be complemented by a brief outline of the approaches taken regarding the field at hand (say, programmatic access to the Web of Linked Data without having to know SPARQL).

Since both BASIL and LD Fragments are incorporated into this work, it might be useful to refer to them and briefly build upon their state of the art.
