### 2021-03-26 CLARIAH/grlc#326 @alexdma
[Software paper] Minor corrections
Re openjournals/joss-reviews#2731

https://github.com/openjournals/joss-papers/blob/joss.02731/joss.02731/10.21105.joss.02731.pdf

can you please check the following

- First paragraph, end: "requires certain level of expertise" -> "a certain level" (or rephrase e.g. "some degree of expertise", "some expertise")
- Second pagaraph: "face the challenge _of_ having to learn SPARQL"
