### 2021-03-26 CLARIAH/grlc#327 @c-martinez
Hi @alexdma! Thanks for raising this issue. Very good point, definitely something to improve there.
