### 2021-04-06 86e1ee7
Update SPARQLTransformer to 2.1.1
