+++ requirements.txt
@@ -17,7 +17,7 @@ requests==2.20.0
 six==1.12.0
 simplejson==3.16.0
 setuptools>=38.6.0
-SPARQLTransformer==1.9.0
+SPARQLTransformer==2.1.1
 SPARQLWrapper==1.8.2
 werkzeug>=0.16.0
 PyGithub==1.43.5
