### 2021-04-15 CLARIAH/grlc#325 @c-martinez
Thanks for your suggestions @alexdma

@albertmeronyo -- would [this](https://github.com/basilapi/basil/wiki/SPARQL-variable-name-convention-for-WEB-API-parameters-mapping) be a good reference for BASIL? And [this one](https://linkeddatafragments.org/publications/jws2016.pdf) for LD fragments?
