+++ requirements.txt
@@ -1,4 +1,5 @@
 docopt==0.6.2
+docutils==0.17.1
 Flask==1.0.2
 Flask-Cors==3.0.6
 gevent==1.4.0
