### 2021-05-06 CLARIAH/grlc#326 @c-martinez
Hi @alexdma -- we've implemented these minor corrections. Thank you for suggesting them!
