+++ requirements.txt
@@ -6,7 +6,7 @@ gevent==1.4.0
 greenlet==0.4.15
 html5lib==1.0.1
 isodate==0.5.4
-itsdangerous==1.1.0
+itsdangerous>=1.1.0
 keepalive==0.5
 MarkupSafe==0.23
 pyaml==18.11.0
