### 2021-05-20 8b2a3d4
Update requirements.txt
itsdangerous is already installed via Flask and Flask-Cors, so no need to include it. Also the fixed version is preventing compatibility with any environment that is working with higher versions
