### 2021-06-22 CLARIAH/grlc#325 @alexdma
Thanks! Where can I view the updated paper?
