### 2021-06-22 CLARIAH/grlc#325 @c-martinez
No problem -- I just generated a [fresh version](https://github.com/openjournals/joss-papers/blob/joss.02731/joss.02731/10.21105.joss.02731.pdf) in the [joss issue](https://github.com/openjournals/joss-reviews/issues/2731)
