### 2021-07-26 f1cdb46
Change warnings into debug messages. See #350
