### 2021-07-26 514cde0
Change warnings into debug messages. See #350
