### 2021-08-03 83b879f
Propagate exceptions raised by rdflib parsing
