### 2021-08-08 CLARIAH/grlc#356 @essepuntato
Comments for submission to JOSS 2731
Some suggestions and comments for improving the software and its description - related to https://github.com/openjournals/joss-reviews/issues/2731.

# License
The license in the repository list only one person as owning the copyright of the software, while there are two authors (and main contributors) listed in the paper describing the software. Just asking but: does it make sense to extend the people listed in the license with both the authors?

# Installation
I've tried to install the software using python 3.8 and python 3.9. The former did not give me any issue. The latter, instead, did not work as expected – see the following screenshot. OS used: macOS 11.5.

# Article
There are some issues in the article presenting the software:

* At the very beginning you mention that RDF is also referred to as Linked Open Data (LOD), but this is not correct. RDF is a tool that enables LOD, but they are not the same thing.

* There is a very limited description of the state of the field, which list only two approaches, but there are surely more - including BASIL, which is cited for a different reason, and OBA. I think would be worth expanding a bit the description of this aspect, including more software and explaining common and different features between grlc and the others. Of course, this impact also the part about the references, which should be expanded accordingly.
