### 2021-08-09 CLARIAH/grlc#356 @c-martinez
Hi @essepuntato -- thank you for your comments, we will take them into account.
