### 2021-09-23 365e24d
Add github action workflows and remove travis config
