### 2021-09-23 afecfa6
Add flake8 testing
