+++ paper.bib
@@ -1,3 +1,17 @@
+@article{Espinoza2021,
+title = {Crossing the chasm between ontology engineering and application development: A survey},
+journal = {Journal of Web Semantics},
+volume = {70},
+pages = {100655},
+year = {2021},
+issn = {1570-8268},
+doi = {https://doi.org/10.1016/j.websem.2021.100655},
+url = {https://www.sciencedirect.com/science/article/pii/S1570826821000305},
+author = {Paola Espinoza-Arias and Daniel Garijo and Oscar Corcho},
+keywords = {Ontology, OWL, Ontology engineering, Web API, Application development, Knowledge graph},
+abstract = {The adoption of Knowledge Graphs (KGs) by public and private organizations to integrate and publish data has increased in recent years. Ontologies play a crucial role in providing the structure for KGs, but are usually disregarded when designing Application Programming Interfaces (APIs) to enable browsing KGs in a developer-friendly manner. In this paper we provide a systematic review of the state of the art on existing approaches to ease access to ontology-based KG data}
+}
+
 @InProceedings{Merono_ISWC2019,
 author="Lisena, Pasquale
 and Mero{\~{n}}o-Pe{\~{n}}uela, Albert
