### 2021-10-20 #2731 @c-martinez
Hi @arfon -- it took a bit longer than expected, but we've now addressed the last of @essepuntato 's comments in https://github.com/CLARIAH/grlc/issues/356. Thanks for all the useful feedback!
