### 2021-10-21 #2731 @arfon
Got it. Thanks @c-martinez. It looks like there are still a couple of issues open on your repo related to the review (e.g., https://github.com/CLARIAH/grlc/issues/311 and https://github.com/CLARIAH/grlc/issues/327). Could you share what your intent here is?

:wave: @essepuntato, @alexdma – could you take a look at the changes that @c-martinez has made and see if you are able to update the remaining checkboxes for the review?
