### 2021-11-02 CLARIAH/grlc#311 @arfon
> I agree that it does. Thank you and nice work adding LDF documentation!

:+1: thanks @alexdma! When you have a moment, please update the review checklist in https://github.com/openjournals/joss-reviews/issues/2731 based on these changes.
