### 2021-11-02 CLARIAH/grlc#311 @alexdma
Closing in agreement with editor in chief
