+++ CITATION.cff
@@ -19,12 +19,11 @@ authors:
     family-names: Zijdeman
     given-names: Richard
 cff-version: "1.0.3"
-date-released: 2017-11-22
-doi: 10.5281/zenodo.1064392
+doi: 10.5281/zenodo.1064391
 license: MIT
 message: "If you use this software, please cite it as below."
 repository-code: "https://github.com/CLARIAH/grlc"
-title: grlc
+title: "grlc: the git repository linked data API constructor"
 abstract: grlc, the git repository linked data API constructor, automatically builds Web APIs using SPARQL queries stored in git repositories.
 keywords:
   - "swagger-ui"
@@ -32,4 +31,5 @@ keywords:
   - "linked-data"
   - "semantic-web"
   - "linked-data-api"
-version: "1.3.6"
+version: "1.3.7"
+date-released: 2021-11-03
