### 2021-11-03 #2731 @c-martinez
Great, thanks @arfon !

I've just made a [release in Zenodo](https://zenodo.org/record/5644276)

I hope everything is as expected?
