### 2021-11-04 #2731 @arfon
@c-martinez - could you please check if those DOIs recommended by Whedon are correct, and if so, add them to your BibTeX file? Also, this one `https://doi.org/10.1016/j.websem.2021.100655` should be written as `10.1016/j.websem.2021.100655` in your BibTeX file please.
