### 2021-11-04 #2731 @arfon
> Sure no problem -- is a new release necessary after fixing those?

No, we can just stick with the same release number and Zenodo archive.
