### 2021-11-04 d4ddb15
Added missing DOIs to paper.bib
