+++ paper.bib
@@ -5,7 +5,7 @@ volume = {70},
 pages = {100655},
 year = {2021},
 issn = {1570-8268},
-doi = {https://doi.org/10.1016/j.websem.2021.100655},
+doi = {10.1016/j.websem.2021.100655},
 url = {https://www.sciencedirect.com/science/article/pii/S1570826821000305},
 author = {Paola Espinoza-Arias and Daniel Garijo and Oscar Corcho},
 keywords = {Ontology, OWL, Ontology engineering, Web API, Application development, Knowledge graph},
@@ -33,7 +33,8 @@ publisher="Springer International Publishing",
 address="Cham",
 pages="454--470",
 abstract="In a document-based world as the one of Web APIs, the triple-based output of SPARQL endpoints can be a barrier for developers who want to integrate Linked Data in their applications. A different JSON output can be obtained with SPARQL Transformer, which relies on a single JSON object for defining which data should be extracted from the endpoint and which shape should they assume. We propose a new approach that amounts to merge SPARQL bindings on the base of identifiers and the integration in the grlc API framework to create new bridges between the Web of Data and the Web of applications.",
-isbn="978-3-030-30796-7"
+isbn="978-3-030-30796-7",
+doi="10.1007/978-3-030-30796-7_28"
 }
 
 @phdthesis{Singh2019,
@@ -41,7 +42,9 @@ isbn="978-3-030-30796-7"
   url = {https://doi.org/10.18174/505685},
   publisher = {Wageningen University and Research},
   author = {Gurnoor Singh},
-  title = {Genomics data integration for knowledge discovery using genome annotations from molecular databases and scientific literature}
+  title = {Genomics data integration for knowledge discovery using genome annotations from molecular databases and scientific literature},
+  institution = "Wageningen University",
+  year = "2019"
 }
 
 @InProceedings{Merono_ISWC2016,
@@ -60,7 +63,8 @@ publisher="Springer International Publishing",
 address="Cham",
 pages="342--353",
 abstract="Building Web APIs on top of SPARQL endpoints is becoming common practice. It enables universal access to the integration favorable data space of Linked Data. In the majority of use cases, users cannot be expected to learn SPARQL to query this data space. Web APIs are the most common way to enable programmatic access to data on the Web. However, the implementation of Web APIs around Linked Data is often a tedious and repetitive process. Recent work speeds up this Linked Data API construction by wrapping it around SPARQL queries, which carry out the API functionality under the hood. Inspired by this development, in this paper we present grlc, a lightweight server that takes SPARQL queries curated in GitHub repositories, and translates them to Linked Data APIs on the fly.",
-isbn="978-3-319-47602-5"
+isbn="978-3-319-47602-5",
+doi="10.1007/978-3-319-47602-5_48"
 }
 
 @InProceedings{Merono_ESWC2017,
@@ -79,7 +83,8 @@ publisher="Springer International Publishing",
 address="Cham",
 pages="143--148",
 abstract="In this demo, we show how an effective and application agnostic way of curating SPARQL queries can be achieved by leveraging Git-based architectures. Often, SPARQL queries are hard-coded into Linked Data consuming applications. This tight coupling poses issues in code maintainability, since these queries are prone to change to adapt to new situations; and query reuse, since queries that might be useful in other applications remain inaccessible. In order to enable decoupling, version control, availability and accessibility of SPARQL queries, we propose SPARQL2Git, an interface for editing, curating and storing SPARQL queries that uses cloud based Git repositories (such as GitHub) as a backend. We describe the query management capabilities of SPARQL2Git, its convenience for SPARQL users that lack Git knowledge, and its combination with grlc to easily generate Linked Data APIs.",
-isbn="978-3-319-70407-4"
+isbn="978-3-319-70407-4",
[…]
